requirements.txt
twitchio
google-generativeai
python-dotenv
aiohttp
